requirements.txt
python-dotenv
schedule
praw
openai
requests
